AMAZON_Measurement_Problems
Business Problem [EN]  One of the most important problems in e-commerce is the correct calculation of the points given to the products after the sale.The solution to this problem is to provide more customer satisfaction for the e-commerce site, the prominence of the product for sellers and a hassle-free shopping for buyers it means experience. Another problem is the correct ordering of the comments given to the products. Since the prominence of misleading comments will directly affect the sale of the product, it will cause both financial loss and customer loss. In solving these 2 basic problems, the e-commerce site and sellers will increase their sales, while the customers will complete the purchase journey smoothly.